+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 3.0.0 (March 11th, 2021)
+
+### Additions
+
+* Support for graph based models. This paves the way for representing more advanced features, such as sub-queries in a more clear way. Since graph based models are similar to DSL models, the hope is that they will be easier to work with and understand.
+
+### Breaking Changes
+
+* The `to_model` method on `Dbee::Base` objects has been removed. Use `to_schema` instead.
+* The `ancestors!` method on `Dbee::Model` has been removed. Use `Dbee::Schema#expand_query_path` instead.
+
 # 2.1.1 (July 14th, 2020)
 
 * Removed guard that ensured a query has at least one field to establish a more rational base-case.